[feat]:confirm

+++ server/src/services/access.service.js
@@ -2,6 +2,7 @@
 
 const crypto = require("crypto");
 const bcrypt = require("bcrypt");
+const process = require("process");
 const KeyTokenService = require("./keyToken.service");
 const { createTokenPair } = require("../auth/authUtils");
 const {
@@ -13,7 +14,10 @@ const { User } = require("../models/user.model");
 const { findByEmail } = require("../models/repo/user.repo");
 const { getInfoData } = require("../utils");
 const sendEmail = require("../helpers/sendEmail");
-const { resetPasswordForm } = require("../utils/emailExtension");
+const {
+  resetPasswordForm,
+  confirmAccountForm,
+} = require("../utils/emailExtension");
 const RolesUser = {
   USER: "USER",
   ADMIN: "ADMIN",
@@ -26,13 +30,36 @@ class AccessService {
       throw new badRequestError("error user already rigisted");
     }
     const passwordHash = await bcrypt.hash(password, 10);
-    await User.create({
+    const user = await User.create({
       userName,
       email,
       phone,
       password: passwordHash,
       roles: [RolesUser.USER],
+      status: "inActive",
     });
+    const key = crypto.randomBytes(64).toString(`hex`);
+    const tokens = await createTokenPair(
+      {
+        userId: user._id,
+        email: user.email,
+        userName: user.userName,
+        phone: user.phone,
+        type: "confirm",
+      },
+      key
+    );
+    await KeyTokenService.createKeyToken({
+      userId: user._id,
+      key,
+    });
+    const confirmAccountLink = `${process.env.BASE_URL_CLIENT}/confirm-account/${tokens.accessToken}`;
+    const confirmAccountFormContent = confirmAccountForm(confirmAccountLink);
+    await sendEmail(
+      user.email,
+      confirmAccountFormContent.title,
+      confirmAccountFormContent.body
+    );
     return {
       data: null,
     };
@@ -64,6 +91,7 @@ class AccessService {
       userId: userId,
       key,
     });
+
     return {
       user: getInfoData({
         fill: ["_id", "userName", "email", "phone"],
@@ -91,7 +119,7 @@ class AccessService {
       userId: user._id,
       key,
     });
-    const resetLink = `http://localhost:5173/reset-password/${tokens.accessToken}`;
+    const resetLink = `${process.env.BASE_URL_CLIENT}/reset-password/${tokens.accessToken}`;
     const resetPasswordFormContent = resetPasswordForm(resetLink);
     await sendEmail(
       user.email,
@@ -109,6 +137,19 @@ class AccessService {
     const userUpdated = await User.findOneAndUpdate(userExiting._id, {
       ...userExiting,
       password: passwordHash,
+    });
+    if (userUpdated) {
+      await KeyTokenService.removeKeyById(keyStore._id);
+    } else throw badRequestError("reset password faild");
+    return "OK";
+  };
+  static confirmAccount = async (user, keyStore) => {
+    const { userId, email } = user;
+    const userExiting = await findByEmail({ email });
+    if (!userExiting?._id && userId) throw new NotFoundError("Not found User");
+    const userUpdated = await User.findOneAndUpdate(userExiting._id, {
+      ...userExiting,
+      status: "active",
     });
     if (userUpdated) {
       await KeyTokenService.removeKeyById(keyStore._id);
